+++ src/engine/H2DE_object_manager.cpp
@@ -39,7 +39,7 @@ void H2DE_ObjectManager::handleMouseDownEvents(SDL_Event event) {
             if (buttonRect.collides(mousePos)) {
 
                 if (button->buttonObjectData.onMouseDown) {
-                    button->buttonObjectData.onMouseDown(button, button->currentTimelineID);
+                    button->buttonObjectData.onMouseDown({ button, button->currentTimelineID });
                 }
                 
                 mouseDown = button;
@@ -67,7 +67,7 @@ void H2DE_ObjectManager::handleMouseUpEvents(SDL_Event event) {
         const H2DE_LevelRect buttonRect = G::getHitboxRect(mouseDown, hitbox);
 
         if (buttonRect.collides(mousePos)) {
-            mouseDown->buttonObjectData.onMouseUp(mouseDown, mouseDown->currentTimelineID);
+            mouseDown->buttonObjectData.onMouseUp({ mouseDown, mouseDown->currentTimelineID });
             mouseDown = nullptr;
             return;
         }
@@ -108,13 +108,13 @@ void H2DE_ObjectManager::handleHoverEvents(SDL_Event event) {
 
             if (hovered != nullptr) {
                 if (hovered->buttonObjectData.onBlur) {
-                    hovered->buttonObjectData.onBlur(hovered, hovered->currentTimelineID);
+                    hovered->buttonObjectData.onBlur({ hovered, hovered->currentTimelineID });
                 }
             }
 
             hovered = button;
             if (button->buttonObjectData.onHover) {
-                button->buttonObjectData.onHover(button, button->currentTimelineID);
+                button->buttonObjectData.onHover({ button, button->currentTimelineID });
             }
 
             return;
@@ -148,7 +148,7 @@ void H2DE_ObjectManager::handleBlurEvents(SDL_Event event) {
 
     if (!stillHovering) {
         if (hovered->buttonObjectData.onBlur) {
-            hovered->buttonObjectData.onBlur(hovered, hovered->currentTimelineID);
+            hovered->buttonObjectData.onBlur({ hovered, hovered->currentTimelineID });
         }
 
         hovered = nullptr;